Split payload module
