+++ bee-test/src/rand/message/payload/indexation.rs
@@ -1,0 +1,17 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::rand::{bytes::rand_bytes, number::rand_number_range};
+
+use bee_message::payload::indexation::{IndexationPayload, INDEXATION_INDEX_LENGTH_RANGE};
+
+/// Generates a random [`IndexationPayload`].
+pub fn rand_indexation_payload() -> IndexationPayload {
+    let index_range = *INDEXATION_INDEX_LENGTH_RANGE.start() as usize..=*INDEXATION_INDEX_LENGTH_RANGE.end() as usize;
+
+    IndexationPayload::new(
+        rand_bytes(rand_number_range(index_range)),
+        rand_bytes(rand_number_range(0..=255)),
+    )
+    .unwrap()
+}
