+++ bee-test/src/rand/message/payload/salt_declaration.rs
@@ -1,0 +1,25 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::rand::{
+    bytes::{rand_bytes, rand_bytes_array},
+    number::rand_number,
+};
+
+use bee_message::payload::salt_declaration::{Salt, SaltDeclarationPayload};
+
+/// Generates a random [`Salt`].
+pub fn rand_salt() -> Salt {
+    Salt::new(rand_bytes(96), rand_number()).unwrap()
+}
+
+/// Generates a random [`SaltDeclarationPayload`].
+pub fn rand_salt_declaration_payload() -> SaltDeclarationPayload {
+    SaltDeclarationPayload::builder()
+        .with_node_id(rand_number())
+        .with_salt(rand_salt())
+        .with_timestamp(rand_number())
+        .with_signature(rand_bytes_array())
+        .finish()
+        .unwrap()
+}
